Gistda-Sementi-Segmentation/fastapi_requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.0.1
opencv-python==4.8.1.78
scikit-image==0.22.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1